feat: 운영시간 로직 구현
