+++ src/page/StoreRegistration/component/TimePicker/TimePicker.module.scss
@@ -35,67 +35,3 @@
     }
   }
 }
-
-.content {
-  position: absolute;
-  display: flex;
-  border: 1px solid #111111;
-  width: 150px;
-  height: 230px;
-  z-index: 2;
-  top: 25px;
-  left: 50;
-
-  &__hour-list {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 75px;
-    height: 230px;
-    overflow: scroll;
-
-    &::-webkit-scrollbar {
-      display: none;
-    }
-  }
-
-  &__hour-item {
-    text-align: center;
-    width: 100%;
-    padding: 10px;
-    background-color: #ffffff;
-    cursor: pointer;
-
-    &:hover {
-      background-color: #175c8e;
-      color: #ffffff;
-    }
-  }
-
-  &__minute-list {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 75px;
-    height: 230px;
-    list-style-type: none;
-    overflow: scroll;
-
-    &::-webkit-scrollbar {
-      display: none;
-    }
-  }
-
-  &__minute-item {
-    text-align: center;
-    width: 100%;
-    padding: 10px;
-    background-color: #ffffff;
-    cursor: pointer;
-
-    &:hover {
-      background-color: #175c8e;
-      color: #ffffff;
-    }
-  }
-}
